requirements.txt
aws-cdk-lib==2.118.0
aws-cdk.aws-lambda-python-alpha==2.118.0a0
constructs>=10.0.0,<11.0.0
boto==2.49.0
cdk-nag>=2.28.0